abstracted components logic into newly created components
